requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.2
defusedxml==0.7.1
Django==3.2.16
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.10
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
Pillow==9.3.0
pluggy==1.0.0.dev0
py==1.11.0
pycparser==2.22
PyJWT==2.10.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2025.2
requests==2.26.0
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.1
sqlparse==0.5.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.20
